code refactor

+++ blog/src/main/java/com/asis/blog/service/serviceimpl/CommentServiceImpl.java
@@ -1,27 +1,35 @@
 package com.asis.blog.service.serviceimpl;
 
+import com.asis.blog.dto.BlogDto;
 import com.asis.blog.dto.CommentDto;
 import com.asis.blog.entity.Blog;
 import com.asis.blog.entity.Comment;
 import com.asis.blog.exception.CustomException;
+import com.asis.blog.helper.ListRemoveHelper;
 import com.asis.blog.mapper.CommentMapper;
 import com.asis.blog.repository.BlogRepository;
 import com.asis.blog.repository.CommentRepository;
+import com.asis.blog.service.BlogService;
 import com.asis.blog.service.CommentService;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.*;
 
 @Service
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
-    private final BlogRepository blogRepository;
+    private final BlogService blogService;
    private final CommentMapper commentMapper;
+   private final ListRemoveHelper listRemoveHelper;
+   private final BlogRepository blogRepository;
 
-    public CommentServiceImpl(CommentRepository commentRepository, BlogRepository blogRepository, CommentMapper commentMapper) {
+    public CommentServiceImpl(CommentRepository commentRepository, BlogService blogService, CommentMapper commentMapper, ListRemoveHelper listRemoveHelper, BlogRepository blogRepository) {
         this.commentRepository = commentRepository;
-        this.blogRepository = blogRepository;
+        this.blogService = blogService;
         this.commentMapper = commentMapper;
+        this.listRemoveHelper = listRemoveHelper;
+        this.blogRepository = blogRepository;
     }
     @Override
     public List<CommentDto> getAllComment() {
@@ -60,51 +68,40 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public List<CommentDto> addCommentToBlog(Long id, CommentDto commentDto) throws CustomException {
-        Optional<Blog> findById = blogRepository.findById(id);
-        if (!findById.isPresent()) {
-            throw new CustomException("no blog of id : " + id + " to add comment");
-        }
-
+    public List<CommentDto> addCommentToBlog(Long blogId, CommentDto commentDto) throws CustomException {
+        Blog blog = blogService.findBlogById(blogId);
         Comment comment = commentMapper.dtoToEntity(commentDto);
         Comment newComment = commentRepository.save(comment);
-        Blog blog = findById.get();
         List<Comment> blogComments = blog.getComments();
         blogComments.add(newComment);
         blog.setComments(blogComments);
-        Blog saveBlog = blogRepository.save(blog);
-
-        return commentMapper.entitiesToDtos(saveBlog.getComments());
+        BlogDto saveBlogDto = blogService.addBlog(blog);
+        return saveBlogDto.getCommentDtos();
 
     }
     @Override
     public String deleteCommentFromComment(Long id) {
         Comment commentParent = commentRepository.findCommentParent(id);
-
+        System.out.println(commentParent.getId());
         if (commentParent != null) {
 //        comments after deleting child
-            commentParent.setComments(removeFromList(id , commentParent.getComments()));
+            listRemoveHelper.removeFromList(id, commentParent.getComments());
+//            commentParent.setComments(commentList);
             commentRepository.save(commentParent);
-            commentRepository.deleteById(id);
-            return "comment deleted from parent of id : "+id;
         }
         commentRepository.deleteById(id);
         return "comment of id : "+id+" deleted";
     }
 
     @Override
-    public String deleteCommentFromBlog(Long id) {
-        Blog blogParent = blogRepository.findCommentParent(id);
-
-        if(blogParent != null){
-           blogParent.setComments(removeFromList(id , blogParent.getComments()));
-            blogRepository.save(blogParent);
-            blogRepository.deleteById(id);
-            return "comment deleted from parent of id : "+id;
-        }
-
+    @Transactional(rollbackOn = Exception.class)
+    public String deleteCommentFromBlog(Long id) throws CustomException {
+        boolean isDeletedFromParentBlog = blogService.deleteCommentFromParentBlog(id);
+        if(isDeletedFromParentBlog){
         commentRepository.deleteById(id);
         return "comment of id : "+id+" deleted";
+        }
+        throw new CustomException("could not delete comment from blog of id : "+id);
     }
 
     @Override
@@ -121,15 +118,5 @@ public class CommentServiceImpl implements CommentService {
         }
         return null;
     }
-
-    private List<Comment> removeFromList(Long id , List<Comment> comments){
-        for (int i = 0 ; i < comments.size() ; i++){
-            if(comments.get(i).getId() == id){
-                comments.remove(i);
-            }
-        }
-        return comments;
-    }
-
 
 }
